update pra rodar com react native web
