+++ .eslintrc.js
@@ -1,4 +1,7 @@
 module.exports = {
   root: true,
   extends: '@react-native-community',
+  rules: {
+    'prettier/prettier': 0,
+  },
 };
