+++ package.json
@@ -7,12 +7,18 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "web": "react-scripts start",
+    "web:start": "react-scripts start",
+    "web:build": "react-scripts build"
   },
   "dependencies": {
-    "react": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-native": "0.66.3",
-    "react-native-svg": "^12.1.1"
+    "react-native-svg": "^12.1.1",
+    "react-native-web": "^0.19.9",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
@@ -27,5 +33,17 @@
   },
   "jest": {
     "preset": "react-native"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
