+++ src/index.js
@@ -1,0 +1,10 @@
+import { AppRegistry } from 'react-native';
+import App from "./App.js";
+
+const appName = "orgsNavegacao";
+
+AppRegistry.registerComponent(appName, () => App);
+AppRegistry.runApplication(appName, {
+    // Mount the react-native app in the "root" div of index.html
+    rootTag: document.getElementById("root"),
+});
